Cleaned up the states by adding a superstate class to contain the transitions and logging.
