+++ src/main/java/frc/robot/fsm/states/DeadeyeState.java
@@ -12,26 +12,15 @@ import org.usfirst.frc3620.logger.LoggingMaster;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.fsm.StateTransition;
+import frc.robot.fsm.SuperState;
 
-public class DeadeyeState implements IState {
+public class DeadeyeState extends SuperState {
 
-    public final static TaggedLogger logger = LoggingMaster.getLogger(DeadeyeState.class);
-
-    private List<StateTransition> stateTransitions = new ArrayList<>();
-
-    public DeadeyeState() {
-    }
-
-    public void addTransition(StateTransition transition) {
-        if (stateTransitions != null) {
-            stateTransitions.add(transition);
-        }
-    }
+    
 
     @Override
     public void onEnter() {
         // Code to run when entering the Scoring state
-        logger.info("Entering Deadeye State");
     }
 
     @Override
@@ -46,17 +35,5 @@ public class DeadeyeState implements IState {
         // Code to run when exiting the Scoring state
 
 
-    }
-
-    @Override
-    public Optional<IState> nextState() {
-        // Logic to determine the next state
-        for (StateTransition transition : stateTransitions) {
-            if (transition.triggerToState().getAsBoolean()) {
-                logger.info("Transitioning from Deadeye to {}", transition.targetState().getClass().getSimpleName());
-                return Optional.of(transition.targetState());
-            }
-        }
-        return Optional.empty();
     }
 }
