+++ src/main/java/frc/robot/fsm/states/HoardingState.java
@@ -12,32 +12,21 @@ import org.usfirst.frc3620.logger.LoggingMaster;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.fsm.StateTransition;
+import frc.robot.fsm.SuperState;
 
-public class HoardingState implements IState {
+public class HoardingState extends SuperState {
 
-    public final static TaggedLogger logger = LoggingMaster.getLogger(HoardingState.class);
 
-    private List<StateTransition> stateTransitions = new ArrayList<>();
-
-    public HoardingState() {
-    }
-
-    public void addTransition(StateTransition transition) {
-        if (stateTransitions != null) {
-            stateTransitions.add(transition);
-        }
-    }
 
     @Override
     public void onEnter() {
         // Code to run when entering the Scoring state
-        logger.info("Entering Hoarding State");
+    
     }
 
     @Override
     public void execute() {
         // Code to run while in the Scoring state
-
 
     }
 
@@ -48,15 +37,5 @@ public class HoardingState implements IState {
 
     }
 
-    @Override
-    public Optional<IState> nextState() {
-        // Logic to determine the next state
-        for (StateTransition transition : stateTransitions) {
-            if (transition.triggerToState().getAsBoolean()) {
-                logger.info("Transitioning from Hoarding to {}", transition.targetState().getClass().getSimpleName());
-                return Optional.of(transition.targetState());
-            }
-        }
-        return Optional.empty();
-    }
+
 }
