+++ src/main/java/frc/robot/fsm/states/ScoringState.java
@@ -12,26 +12,15 @@ import org.usfirst.frc3620.logger.LoggingMaster;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
 import frc.robot.fsm.StateTransition;
+import frc.robot.fsm.SuperState;
 
-public class ScoringState implements IState {
+public class ScoringState extends SuperState {
 
-    public final static TaggedLogger logger = LoggingMaster.getLogger(ScoringState.class);
 
-    private List<StateTransition> stateTransitions = new ArrayList<>();
-
-    public ScoringState() {
-    }
-
-    public void addTransition(StateTransition transition) {
-        if (stateTransitions != null) {
-            stateTransitions.add(transition);
-        }
-    }
 
     @Override
     public void onEnter() {
         // Code to run when entering the Scoring state
-        logger.info("Entering Scoring State");
     }
 
     @Override
@@ -48,15 +37,4 @@ public class ScoringState implements IState {
 
     }
 
-    @Override
-    public Optional<IState> nextState() {
-        // Logic to determine the next state
-        for (StateTransition transition : stateTransitions) {
-            if (transition.triggerToState().getAsBoolean()) {
-                logger.info("Transitioning from Scoring to {}", transition.targetState().getClass().getSimpleName());
-                return Optional.of(transition.targetState());
-            }
-        }
-        return Optional.empty();
-    }
 }
